fix: style fixes, moved function to utils
